Move refinery and harvester logic to Mods.Common;
Remove unused interface IAcceptOreDockAction
