+++ OpenRA.Mods.Common/Activities/DeliverResources.cs
@@ -10,12 +10,10 @@
 
 using System.Drawing;
 using OpenRA.Activities;
-using OpenRA.Mods.Common.Activities;
 using OpenRA.Mods.Common.Traits;
-using OpenRA.Mods.RA.Traits;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.Activities
+namespace OpenRA.Mods.Common.Activities
 {
 	public class DeliverResources : Activity
 	{
