+++ OpenRA.Mods.Common/Activities/HarvesterDockSequence.cs
@@ -11,11 +11,10 @@
 using System;
 using System.Collections.Generic;
 using OpenRA.Activities;
-using OpenRA.Mods.Common.Activities;
-using OpenRA.Mods.RA.Traits;
+using OpenRA.Mods.Common.Traits;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.Activities
+namespace OpenRA.Mods.Common.Activities
 {
 	public class HarvesterDockSequence : Activity
 	{
