+++ OpenRA.Mods.Common/Traits/Player/HarvesterAttackNotifier.cs
@@ -9,10 +9,9 @@
 #endregion
 
 using System.Drawing;
-using OpenRA.Mods.Common.Traits;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.Traits
+namespace OpenRA.Mods.Common.Traits
 {
 	[Desc("Plays an audio notification and shows a radar ping when a harvester is attacked.",
 		"Attach this to the player actor.")]
