+++ OpenRA.Mods.Common/Traits/Render/RenderHarvester.cs
@@ -10,10 +10,9 @@
 
 using OpenRA.Activities;
 using OpenRA.Graphics;
-using OpenRA.Mods.Common.Traits;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.Traits
+namespace OpenRA.Mods.Common.Traits
 {
 	class RenderHarvesterInfo : RenderUnitInfo, Requires<HarvesterInfo>
 	{
