+++ OpenRA.Mods.RA/Traits/HarvesterHuskModifier.cs
@@ -8,7 +8,7 @@
  */
 #endregion
 
-using OpenRA.Mods.RA;
+using OpenRA.Mods.Common.Traits;
 using OpenRA.Traits;
 
 namespace OpenRA.Mods.RA.Traits
