--- OpenRA.Mods.RA/TraitsInterfaces.cs
@@ -1,28 +1,0 @@
-#region Copyright & License Information
-/*
- * Copyright 2007-2015 The OpenRA Developers (see AUTHORS)
- * This file is part of OpenRA, which is free software. It is made
- * available to you under the terms of the GNU General Public License
- * as published by the Free Software Foundation. For more information,
- * see COPYING.
- */
-#endregion
-
-using OpenRA.Mods.RA.Activities;
-
-namespace OpenRA.Mods.RA
-{
-	public interface IAcceptResources
-	{
-		void OnDock(Actor harv, DeliverResources dockOrder);
-		void GiveResource(int amount);
-		bool CanGiveResource(int amount);
-		CVec DeliveryOffset { get; }
-		bool AllowDocking { get; }
-	}
-
-	public interface IAcceptOreDockAction
-	{
-		void OnDock(Actor self, Actor harv, DeliverResources dockOrder);
-	}
-}
